+++ src/test/php/FunctionsTest.php
@@ -8,8 +8,6 @@
  * @package  stubbles\webapp\session
  */
 namespace stubbles\webapp\session;
-use stubbles\input\ValueReader;
-use stubbles\ioc\Binder;
 /**
  * Tests for stubbles\webapp\session\*().
  *
@@ -23,14 +21,9 @@ class FunctionsTest extends \PHPUnit_Framework_TestCase
      */
     public function nativeCreatesWebSession()
     {
-        $sessionCreator = native('example');
-        $mockRequest    = $this->getMock('stubbles\input\web\WebRequest');
-        $mockRequest->expects($this->once())
-                    ->method('readHeader')
-                    ->will($this->returnValue(ValueReader::forValue('example user agent')));
         $this->assertInstanceOf(
                 'stubbles\webapp\session\WebSession',
-                 $sessionCreator($mockRequest)
+                 native('example', md5('example user agent'))
         );
     }
 
@@ -39,43 +32,9 @@ class FunctionsTest extends \PHPUnit_Framework_TestCase
      */
     public function noneDurableCreatesNullSession()
     {
-        $sessionCreator = noneDurable();
         $this->assertInstanceOf(
                 'stubbles\webapp\session\NullSession',
-                 $sessionCreator()
+                 noneDurable()
         );
-    }
-
-    /**
-     * @test
-     */
-    public function bindBindsSessionToPassedInstance()
-    {
-        $binder = new Binder();
-        $mockSession = $this->getMock('stubbles\webapp\session\Session');
-        bind($binder, $mockSession);
-        $this->assertSame(
-                $mockSession,
-                $binder->getInjector()->getInstance('stubbles\webapp\session\Session')
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function bindInitializesSessionScope()
-    {
-        $binder = new Binder();
-        $mockSession = $this->getMock('stubbles\webapp\session\Session');
-        bind($binder, $mockSession);
-        try {
-            $binder->bind('\stdClass')
-                   ->to('\stdClass')
-                   ->inSession();
-        } catch (RuntimeException $re) {
-            $this->fail($re->getMessage());
-        }
-
-        $this->addToAssertionCount(1);
     }
 }