+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+5.0.0 (2014-08-??)
+------------------
+
+### BC breaks
+
+  * session instance creation functions `stubbles\webapp\session\native()` and `stubbles\webapp\session\noneDurable()` now return the session instance directly instead of a callable
+  * removed `stubbles\webapp\session\bind()`
+  * removed `stubbles\webapp\session\ioc\SessionBindingScope`
+
+
+### Other changes
+
+   * removed dependency to stubbles/core
+   * removed dependency to stubbles/input
+
+
 4.0.0 (2014-07-31)
 ------------------